Working on using axios
I'm having trouble getting the information I want out of the promises.

In the console, I get a promise with a promise value of the author object I want but I don't know how I'm suppose to get this data. I tried setting reponse.data to a variable to return but the promise reolves after the method returns.

I'm going to continue to research this.